refactor: move lib/constant under view/*/components
